configurable placementof  imports without preexisting package matches
if the package of the import isn't found in the present imports, allow
the variable pymport_package_precedence to specify which packages are to
be placed below any other imports.
if the target is found in that list, its position in and the order of
the list is used to determine the location of the new import. if it
isn't found, place it above any import of packages of that list.